requirements.txt
azure-cognitiveservices-speech
python-dotenv
fastapi
uvicorn
streamlit
requests
pandas
streamlit-audiorec